Renamed analysis to workflow
