--- cg/cli/analysis/__init__.py
@@ -1,7 +1,0 @@
-"""Adds cg cli analysis command and sub commands"""
-
-from cg.cli.analysis.analysis import analysis
-from cg.cli.analysis.balsamic import balsamic
-from cg.cli.analysis.microsalt import microsalt
-from cg.cli.analysis.mip_rd_dna import mip_dna
-from cg.cli.analysis.mip_rd_rna import mip_rna
