+++ cg/cli/workflow/mip_rd_rna.py
@@ -8,7 +8,7 @@ from cg.apps import hk, tb, scoutapi, lims
 from cg.apps.environ import environ_email
 from cg.apps.mip import MipAPI
 from cg.apps.mip.fastq import MipFastqHandler
-from cg.cli.analysis.analysis import get_links
+from cg.cli.workflow.workflow import get_links
 from cg.meta.analysis import AnalysisAPI
 from cg.meta.deliver.mip_rna import MipRnaDeliverAPI as DeliverAPI
 from cg.store import Store
