+++ tests/cli/workflow/balsamic/test_cli_balsamic_config_case.py
@@ -3,7 +3,7 @@ import logging
 from pathlib import Path
 
 import pytest
-from cg.cli.analysis.balsamic import config_case
+from cg.cli.workflow.balsamic import config_case
 
 EXIT_SUCCESS = 0
