+++ tests/cli/workflow/balsamic/test_cli_balsamic_start.py
@@ -1,7 +1,7 @@
 """This script tests the cli methods to start run balsamic"""
 import logging
 
-from cg.cli.analysis.balsamic import start
+from cg.cli.workflow.balsamic import start
 
 EXIT_SUCCESS = 0
